+++ database/migrations/2018_09_24_131400_create_facebook_account_group_pivot_table.php
@@ -16,7 +16,7 @@ class CreateFacebookAccountGroupPivotTable extends Migration
         Schema::create('facebook_account_group', function (Blueprint $table) {
             $table->bigInteger('facebook_account_id')->unsigned()->index();
             $table->foreign('facebook_account_id')->references('id')->on('facebook_accounts')->onDelete('cascade');
-            $table->integer('group_id')->unsigned()->index();
+            $table->bigInteger('group_id')->unsigned()->index();
             $table->foreign('group_id')->references('id')->on('groups')->onDelete('cascade');
             $table->primary(['facebook_account_id', 'group_id']);
         });
